NoteCell separated
